+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [0.0.79] - 2025-02-25
 
+- [#59](https://github.com/estruyf/vscode-demo-time/issues/59): Added slide controls next and previous (if enabled)
 - [#60](https://github.com/estruyf/vscode-demo-time/issues/60): Cursor visibility on the slide view
 
 ## [0.0.78] - 2025-02-25